+++ src/data/fetchData.js
@@ -319,135 +319,3 @@ export function addMissing(rows, sensor) {
   });
   return imputedRows;
 }
-
-/**
- *
- * @param {Promise<EpiDataRow[]>[]} all
- * @param {((i: number) => SensorEntry)} sensorOf
- * @param {Date} startDate
- * @param {Date} endDate
- * @param {boolean} fitRange
- */
-function syncStartEnd(all, sensorOf, startDate, endDate, fitRange = false) {
-  const adaptMinMax = (r, i, min, max) => {
-    if (r.length === 0) {
-      return r;
-    }
-    if (r[0].date_value != null && r[0].date_value.getTime() > min) {
-      // inject a min
-      r.unshift(createCopy(r[0], new Date(min), sensorOf(i)));
-    }
-    if (r[r.length - 1].date_value != null && r[r.length - 1].date_value.getTime() < max) {
-      // inject a max
-      r.push(createCopy(r[r.length - 1], new Date(max), sensorOf(i)));
-    }
-    return r;
-  };
-  if (fitRange) {
-    // fast can adapt every one independently
-    return all.map((promise, i) => promise.then((r) => adaptMinMax(r, i, startDate.getTime(), endDate.getTime())));
-  }
-  // sync start and end date
-  const allDone = Promise.all(all).then((rows) => {
-    const min = rows.reduce(
-      (acc, r) => (r.length === 0 || r[0].date_value == null ? acc : Math.min(acc, r[0].date_value.getTime())),
-      Number.POSITIVE_INFINITY,
-    );
-    const max = rows.reduce(
-      (acc, r) =>
-        r.length === 0 || r[r.length - 1].date_value == null
-          ? acc
-          : Math.max(acc, r[r.length - 1].date_value.getTime()),
-      Number.NEGATIVE_INFINITY,
-    );
-    return rows.map((r, i) => adaptMinMax(r, i, min, max));
-  });
-  return all.map((_, i) => allDone.then((r) => r[i]));
-}
-
-/**
- * fetch multiple signals and synchronizes their start / end date
- * @param {SensorEntry[]} sensorEntries
- * @param {string} level
- * @param {string | undefined} region
- * @param {Date} startDate
- * @param {Date} endDate
- * @param {boolean} fitRange
- * @param {Partial<EpiDataRow>} mixinValues
- * @returns {Promise<EpiDataRow[]>[]}
- */
-export function fetchMultipleTimeSlices(
-  sensorEntries,
-  level,
-  region,
-  startDate = START_TIME_RANGE,
-  endDate = END_TIME_RANGE,
-  fitRange = false,
-  mixinValues = {},
-) {
-  const all = sensorEntries.map((entry) =>
-    fetchTimeSlice(entry, level, region, startDate, endDate, false, mixinValues),
-  );
-  if (sensorEntries.length <= 1) {
-    return all;
-  }
-  return syncStartEnd(all, (i) => sensorEntries[i], startDate, endDate, fitRange);
-}
-
-/**
- * fetch multiple regions and synchronizes their start / end date
- * @param {SensorEntry} sensorEntry
- * @param {string} level
- * @param {string[]} region
- * @param {Date} startDate
- * @param {Date} endDate
- * @param {boolean} fitRange
- * @param {Partial<EpiDataRow>} mixinValues
- * @returns {Promise<EpiDataRow[]>[]}
- */
-export function fetchMultipleRegionsTimeSlices(
-  sensorEntry,
-  level,
-  regions,
-  startDate = START_TIME_RANGE,
-  endDate = END_TIME_RANGE,
-  fitRange = false,
-  mixinValues = {},
-) {
-  const all = regions.map((region) =>
-    fetchTimeSlice(sensorEntry, level, region, startDate, endDate, false, mixinValues),
-  );
-  if (regions.length <= 1) {
-    return all;
-  }
-  return syncStartEnd(all, () => sensorEntry, startDate, endDate, fitRange);
-}
-
-/**
- * fetches data for a specific data and region for multiple signals in the same data source
- * @param {string} dataSource
- * @param {string[]} signals
- * @param {Date | Date[]} date
- * @param {import('../maps').NameInfo} region
- * @param {string[]} extraFields
- * @returns {Promise<EpiDataRow[]>[]}
- */
-export function fetchMultiSignal(dataSource, signals, date, region, extraFields) {
-  const mixinValues = {
-    geo_value: region.propertyId,
-  };
-  if (!Array.isArray(date)) {
-    mixinValues.time_value = formatAPITime(date);
-  }
-  const transferFields = [...computeTransferFields(mixinValues), ...extraFields];
-
-  return callAPIEndPoint(
-    undefined,
-    dataSource,
-    signals.join(','),
-    region.level,
-    Array.isArray(date) ? date.map((d) => formatAPITime(d)).join(',') : date,
-    region.propertyId,
-    [...transferFields, 'signal'],
-  ).then((rows) => parseData(rows, mixinValues));
-}
