--- src/data/screenshot.js
@@ -1,11 +1,0 @@
-export function downloadUrl(url, name) {
-  const a = document.createElement('a');
-  a.download = name;
-  a.href = url;
-  a.style.position = 'absolute';
-  a.style.left = '-10000px';
-  a.style.top = '-10000px';
-  document.body.appendChild(a);
-  a.click();
-  a.remove();
-}
