+++ src/util.js
@@ -85,3 +85,15 @@ export function unobserveResize(element) {
   observerListeners.delete(element);
   observer.unobserve(element);
 }
+
+export function downloadUrl(url, name) {
+  const a = document.createElement('a');
+  a.download = name;
+  a.href = url;
+  a.style.position = 'absolute';
+  a.style.left = '-10000px';
+  a.style.top = '-10000px';
+  document.body.appendChild(a);
+  a.click();
+  a.remove();
+}
